requirements.txt
Flask==2.3.3
bootstrap-flask==2.5.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
loguru==0.7.2.0
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.2
pytest-cov==4.1.0
email-validator==2.1.0
WTForms==3.0.1